Updated chrome image / alpine-chrome (fork)
I have switched to another image which is regular updated.
The security is also improved in a more recent version.

image: ghcr.io/zoobdude/alpine-chrome:latest

https://github.com/Zoobdude/alpine-chrome

alpine-chrome (fork)

Chrome running in headless mode in a tiny Alpine image

Thank you to the original alpine-chrome team for all the heavy lifting! This repository simply continues regular builds of the original project (with more recent Chrome versions) as the original is no longer maintained.

It would be nice if you accept the change.

Thank you.